move palette to visualizer
