+++ mmdet/visualization/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .local_visualizer import DetLocalVisualizer
-from .palette import get_palette, palette_val
+from .local_visualizer import DetLocalVisualizer, get_palette, palette_val
 
-__all__ = ['palette_val', 'get_palette', 'DetLocalVisualizer']
+__all__ = ['DetLocalVisualizer', 'palette_val', 'get_palette']
